Added enemy wave spawning
Enemies are spawned in via a Spawner object and reference to them is stored in a currentWave List.
However after they died, their references would never clear. The SpawnerBehaviour could never be alerted of an enemies' death outside of using events. Therefore EnemyWaveLink is a component that lets the enemy hold a reference to the WaveSpawner, rather than the other way round. This allows the enemy to remove themselves from the Wave List before they die