Fixed a bug that made bullets randomly move in the world angle 0,0,0
FINALLY this thing was slightly annoying me fo ages and I realised that it only happens to a few bullet types, even though they all shared the same code. So I concluded it was the execution order of lines in scripts.
The bullet would spawn (pointing at 0,0,0 default-ly) and sometimes the ApplyForce would happen before it's transform.forward was set to aimDirection, hence it going off in the direction 0,0,0.
I made it so the bullet always instantiates in aimDirection's direction, meaning all tha shi happens in one line, instantly.