Revamped weapon behaviour
All weapon types now connect to a ShootBehaviour component, each with a different version of the Fire() method, allowing any type of weapon to shoot in any type of way.
Haven't taken a look at Lightning or Ice. Plus Enemies have a bug where they are using the default ShooterBehaviour version of CalculateAimDirection, which uses player inputs. Resulting in the enemy shooting wherever the player is looking.