Added aiming function Component, to have custom aiming logic
This does fix the Enemy shooting in the wrong direction, but I just kinda hate this idea. I've added another component you gotta make sure to have on a weapon for it to work. This logic should be stored in each individual ShootBehaviour subclass (like it was). I'm probably going to backtrack to find a better solution somewhen.